Refactored JWS-specific logic out of PyJWT and into PyJWS superclass
