+++ jwt/__init__.py
@@ -16,7 +16,7 @@ __license__ = 'MIT'
 __copyright__ = 'Copyright 2015 José Padilla'
 
 
-from .api import (
+from .api_jwt import (
     encode, decode, register_algorithm, unregister_algorithm, PyJWT
 )
 from .exceptions import (
